BridgeTestReporter - move to Tester package on Android as well
