+++ native/android/src/main/java/com/nozbe/watermelondb/WatermelonDBPackage.kt
@@ -5,12 +5,11 @@ import com.facebook.react.bridge.JavaScriptModule
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
-import com.nozbe.watermelondb.helpers.BridgeTestReporter
 
 class WatermelonDBPackage : ReactPackage {
 
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> =
-        listOf(DatabaseBridge(reactContext), BridgeTestReporter(reactContext))
+        listOf(DatabaseBridge(reactContext))
 
     @Deprecated("Deprecated RN 0.47", ReplaceWith("createViewManagers()"))
     fun createJSModules(): List<Class<out JavaScriptModule>> = emptyList()
