+++ native/androidTest/app/src/main/java/com/nozbe/watermelonTest/NativeModulesPackage.kt
@@ -5,11 +5,13 @@ import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 
+import com.nozbe.watermelonTest.BridgeTestReporter
+
 class NativeModulesPackage : ReactPackage {
 
     override fun createViewManagers(reactContext: ReactApplicationContext)
             : List<ViewManager<*, *>> = emptyList()
 
     override fun createNativeModules(reactContext: ReactApplicationContext)
-            : List<NativeModule> = emptyList()
+            : List<NativeModule> = listOf(BridgeTestReporter(reactContext))
 }
